feat: use distroless/static as base for docker images
BREAKING CHANGE: Instead of alpine the docker images are now based on
distroless-static. This should minimize attack surface of the exporter
container, but comes at the cost of not having a shell in the
container. The exporter process now also runs as a non-root user to further
minimize attack surface.